requirements.txt
annotated-types==0.6.0
APScheduler==3.10.4
asgiref==3.7.2
Babel==2.13.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
croniter==2.0.1
Flask==2.3.3
Flask-APScheduler==1.13.1
flask-babel==4.0.0
Flask-WTF==1.2.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.2
Pillow==9.5.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.6
webargs==8.3.0
Werkzeug==2.3.7
WTForms==3.0.1
